// back_office/src/app/shared/enums/authorities.enum.ts
export enum Authorities {
  DPM_ADMINISTRATOR = "DPM_ADMINISTRATOR",
  TEST = 'TEST',
  BS_AUDIT = "BS_AUDIT",
  BS_ECTD = "BS_REDIRECTION_ECTD",
  BS_ADMIN = "BS_ADMIN",
  BS_USER = "BS_USER",
  BS_USER_GROUP = 'BS_USER_GROUP',
  BS_TASKS = 'BS_TASKS',
  BS_SUBMIT_ENGAGEMENT_RSE ="BS_SUBMIT_ENGAGEMENT_RSE",
  BS_VIEW_BONNE_PRATIQUE = "BS_VIEW_BONNE_PRATIQUE",
  BS_SAVE_AS_DRAFT_BONNE_PRATIQUE = "BS_SAVE_AS_DRAFT_BONNE_PRATIQUE",
  BS_EDIT_ENGAGEMENT_RSE = "BS_EDIT_ENGAGEMENT_RSE",
  BS_SAVE_AS_DRAFT_ENGAGEMENT_RSE = "BS_SAVE_AS_DRAFT_ENGAGEMENT_RSE",
  BS_SUBMIT_DECLARATION_DEBUT_ACTIVITE = "BS_SUBMIT_DECLARATION_DEBUT_ACTIVITE",
  BS_EDIT_DECLARATION_DEBUT_ACTIVITE = "BS_EDIT_DECLARATION_DEBUT_ACTIVITE",
  BS_SUBMIT_BONNE_PRATIQUE = "BS_SUBMIT_BONNE_PRATIQUE",
  BS_SAVE_AS_DRAFT_DECLARATION_DEBUT_ACTIVITE = "BS_SAVE_AS_DRAFT_DECLARATION_DEBUT_ACTIVITE",
  BS_VIEW_DECLARATION_DEBUT_ACTIVITE = "BS_VIEW_DECLARATION_DEBUT_ACTIVITE",
  BS_VIEW_ENGAGEMENT_RSE = "BS_VIEW_ENGAGEMENT_RSE",
  BS_EDIT_BONNE_PRATIQUE = "BS_EDIT_BONNE_PRATIQUE",
  BS_VIEW_EQUITE_RECRUCTEMENT = "BS_VIEW_EQUITE_RECRUCTEMENT",
  BS_EDIT_EQUITE_RECRUTEMENT = "BS_EDIT_EQUITE_RECRUTEMENT",
  BS_SUBMIT_EQUITE_RECRUTEMENT = "BS_SUBMIT_EQUITE_RECRUTEMENT",
  BS_SAVE_AS_DRAFT_EQUITE_RECRUTEMENT = "BS_SAVE_AS_DRAFT_EQUITE_RECRUTEMENT",
  BS_VIEW_ENGAGEMENT_DIGITAL = "BS_VIEW_ENGAGEMENT_DIGITAL",
  BS_EDIT_ENGAGEMENT_DIGITAL = "BS_EDIT_ENGAGEMENT_DIGITAL",
  BS_SUBMIT_ENGAGEMENT_DIGITAL = "BS_SUBMIT_ENGAGEMENT_DIGITAL",
  BS_SAVE_AS_DRAFT_ENGAGEMENT_DIGITAL = "BS_SAVE_AS_DRAFT_ENGAGEMENT_DIGITAL",
  BS_VIEW_DECLARATION_INTERESTS="BS_VIEW_DECLARATION_INTERESTS",
  BS_WAITLIST_REGISTRATION="BS_WAITLIST_REGISTRATION",
  BS_DESIGNATIONS_USERS="BS_DESIGNATIONS_USERS",
  BS_CREATE_CRO_ACCOUNT='BS_CREATE_CRO_ACCOUNT',
  BS_CREATE_IMPORT_EXPORT='BS_CREATE_IMPORT_EXPORT',
  BS_CREATE_PROMOT_AGENCY='BS_CREATE_PROMOT_AGENCY',
  BS_CREATE_PCT='BS_CREATE_PCT',
  BS_CREATE_AGENCE_PCT='BS_CREATE_AGENCE_PCT',
  BS_CREATE_SUBSIDIARY_PCT='BS_CREATE_SUBSIDIARY_PCT',
  BS_CREATE_PHARMACY='BS_CREATE_PHARMACY',
  BS_OPEN_PHARMACY='BS_OPEN_PHARMACY',
  BS_CREATE_PARAPHARAMACY='BS_CREATE_PARAPHARAMACY',
  BS_CREATE_WHOLESALER='BS_CREATE_WHOLESALER',
  BS_WAITLIST_UNSUBSCRIBE='BS_WAITLIST_UNSUBSCRIBE',
  BS_WAIT_RENEW_REGION_CHANGE='BS_WAIT_RENEW_REGION_CHANGE',
  BS_WAIT_RENEW_CATEGOR_CHANG='BS_WAIT_RENEW_CATEGOR_CHANG',
  BS_TRACK_REQUESTS='BS_TRACK_REQUESTS',
  BS_LABORATORY='BS_LABORATORY',


}